add collection metadata verification
