+++ programs/mmm/src/instructions/cnft/mod.rs
@@ -1,3 +1,5 @@
 pub mod sol_cnft_fulfill_buy;
+pub mod metadata_args;
 
 pub use sol_cnft_fulfill_buy::*;
+pub use metadata_args::*;
