+++ programs/mmm/src/instructions/cnft/sol_cnft_fulfill_buy.rs
@@ -6,71 +6,11 @@ use crate::{
     constants::*,
     errors::MMMErrorCode,
     state::{BubblegumProgram, Pool, SellState, TreeConfigAnchor},
-    util::{log_pool, transfer_compressed_nft, try_close_pool},
+    util::{hash_metadata, log_pool, transfer_compressed_nft, try_close_pool},
     verify_referral::verify_referral,
 };
 
-// Define the TokenStandard enum
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq, PartialOrd, Hash)]
-pub enum TokenStandard {
-    NonFungible,
-    FungibleAsset,
-    Fungible,
-    NonFungibleEdition,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq)]
-pub struct Collection {
-    pub verified: bool,
-    pub key: Pubkey,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq, PartialOrd, Hash)]
-pub enum UseMethod {
-    Burn,
-    Multiple,
-    Single,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq)]
-pub struct Uses {
-    pub use_method: UseMethod,
-    pub remaining: u64,
-    pub total: u64,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq, PartialOrd, Hash)]
-pub enum TokenProgramVersion {
-    Original,
-    Token2022,
-}
-
-#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Eq, PartialEq)]
-pub struct Creator {
-    pub address: Pubkey,
-    pub verified: bool,
-    /// The percentage share.
-    ///
-    /// The value is a percentage, not basis points.
-    pub share: u8,
-}
-
-// Define the MetadataArgs struct
-#[derive(AnchorSerialize, AnchorDeserialize, Clone)]
-pub struct MetadataArgs {
-    pub name: String,
-    pub symbol: String, // Changed from Option<String> to String
-    pub uri: String,
-    pub seller_fee_basis_points: u16,
-    pub primary_sale_happened: bool, // Changed from Option<bool> to bool
-    pub is_mutable: bool,            // Changed from Option<bool> to bool
-    pub edition_nonce: Option<u8>,
-    pub token_standard: Option<TokenStandard>, // Changed from Option<u8> to Option<TokenStandard>
-    pub collection: Option<Collection>,
-    pub uses: Option<Uses>,
-    pub token_program_version: TokenProgramVersion, // Assuming TokenProgramVersion is a simple u8
-    pub creators: Vec<Creator>,
-}
+use super::MetadataArgs;
 
 #[derive(AnchorSerialize, AnchorDeserialize)]
 pub struct SolCnftFulfillBuyArgs {
@@ -210,6 +150,8 @@ pub fn handler<'info>(
         return Err(MMMErrorCode::InvalidAccountState.into());
     }
 
+    let data_hash = hash_metadata(&args.metadata_args)?;
+
     msg!("seller fee basis points: {}", args.seller_fee_basis_points);
     // Create data_hash from metadata_hash + seller_fee_basis_points (secures creator royalties)
     // let data_hash = hash_metadata_data(args.metadata_hash, args.seller_fee_basis_points)?;
@@ -228,7 +170,7 @@ pub fn handler<'info>(
         proof_path,
         ctx.accounts.bubblegum_program.key(),
         args.root,
-        args.metadata_hash,
+        data_hash,
         args.creator_hash,
         args.nonce,
         args.index,
